Correct types for enumeration types and constant values Change the constant values for the protocol in the IPv4 packet to be in an enumeration type

+++ PacketCaptureAnalyser/Ethernet Frame/ARP Packet/ARPPacketConstants.cs
@@ -33,6 +33,6 @@ namespace EthernetFrameNamespace.ARPPacketNamespace
 
         //Length
 
-        public const int ARPPacketHeaderLength = 28;
+        public const ushort ARPPacketHeaderLength = 28;
     }
 }
